Add async task for audio transcription
This commit introduces an asynchronous task using Celery to handle audio file
transcription. The task downloads the audio file, transcribes it using Whisper,
and updates the database with the transcription result.

The previous synchronous implementation caused delays in the user interface,
and this change aims to improve user experience by handling the transcription
in the background.
