+++ transcribe/transcribe/settings.py
@@ -138,3 +138,8 @@ AWS_S3_REGION_NAME = env('AWS_S3_REGION_NAME')
 # For serving static files directly from S3
 AWS_S3_USE_SSL = True
 AWS_S3_VERIFY = True
+
+# Celery settings
+CELERY_BROKER_URL = 'redis://localhost:6379'
+CELERY_RESULT_BACKEND = 'redis://localhost:6379'
+CELERY_BROKER_CONNECTION_RETRY_ON_STARTUP = True
